test(server-rs): differential harness for /api/sites, and the parity fixes it found
Three real differences, all fixed: the shared statement text
`SELECT organization_id FROM sites WHERE site_id = $1` must bind an i64 like
auth does or sqlx's per-connection prepared-statement cache poisons the other
callers; an ISO timestamp bound to a `timestamp` column needs an explicit
::timestamp cast because postgres-js sends it untyped; and the ClickHouse insert
error a failed import batch reports is the client's parsed message, not the raw
response body.

+++ server-rs/src/api/sites/usage.rs
@@ -148,7 +148,10 @@ pub async fn get_site_usage(
 
     let organization_id: Result<Option<Option<String>>, sqlx::Error> =
         sqlx::query_scalar("SELECT organization_id FROM sites WHERE site_id = $1 LIMIT 1")
-            .bind(numeric)
+            // i64 like the identical statement in auth and the analytics chain: sqlx
+            // caches prepared statements per connection by SQL text, and an int4 bound
+            // to the cached int8 one fails
+            .bind(i64::from(numeric))
             .fetch_optional(&state.pg)
             .await;
     let organization_id = match organization_id {
